Refactor des tests unitaires avec PHPUnit
