--- .atoum.php
@@ -1,3 +1,0 @@
-<?php
-
-$runner->addTestsFromDirectory(__DIR__ . '/tests/units');
